[Rename] responseMessage.ts 오타 수정
